backend/showroomapi/requirements.txt
asgiref==3.5.2
astroid==2.12.13
async-timeout==4.0.2
attrs==22.2.0
autobahn==22.12.1
Automat==22.10.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
constantly==15.1.0
cryptography==39.0.0
daphne==4.0.0
dill==0.3.6
Django==4.1.4
django-auto-prefetching==0.2.12
django-cors-headers==3.13.0
django-filter==22.1
django-watchman==1.3.0
django_debug_toolbar==3.8.1
djangochannelsrestframework==1.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-ujson==1.2
drf-ujson2==1.7.2
flake8==5.0.4
flake8-django==1.1.5
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
isort==5.11.2
lazy-object-proxy==1.8.0
mccabe==0.7.0
msgpack==1.0.4
Pillow==9.3.0
platformdirs==2.6.0
psutil==5.9.4
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.6.0
pylint==2.15.8
pyOpenSSL==23.0.0
python-decouple==3.6
pytz==2022.6
pywatchman==1.4.1
redis==4.4.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
ujson==5.7.0
urllib3==1.26.13
urllib3-secure-extra==0.1.0
wrapt==1.14.1
zope.interface==5.5.2
